Move upgrade state logic into upgrade component
