+++ src/frontend/screens/SyncModal/UpgradeBar.js
@@ -1,14 +1,22 @@
+// @flow
 import React from "react";
 import { defineMessages, FormattedMessage, useIntl } from "react-intl";
 import { MaterialIndicator } from "react-native-indicators";
 import { View, StyleSheet } from "react-native";
+import rnBridge from "nodejs-mobile-react-native";
 
 import Text from "../../sharedComponents/Text";
 import { TouchableNativeFeedback } from "../../sharedComponents/Touchables";
 import { DoneIcon, ErrorIcon } from "../../sharedComponents/icons";
-import { UpgradeState } from "./index";
 import Progress from "../../sharedComponents/icons/Progress";
 import DotIndicator from "./DotIndicator";
+import { peerStatus } from "./PeerList";
+import { UpgradeState as BackendUpgradeState } from "../../../backend/lib/constants";
+import ApkInstaller from "../../lib/ApkInstaller";
+import type { Peer } from "./PeerList";
+
+import debug from "debug";
+const log = debug("mapeo-mobile:SyncModal:index");
 
 const m = defineMessages({
   checkingUpdates: {
@@ -72,12 +80,101 @@ const m = defineMessages({
   },
 });
 
-const TextButton = ({ onPress, title }) => (
+const UpgradeState = {
+  Searching: 0,
+  Downloading: 1,
+  GenericError: 2,
+  PermissionError: 3,
+  ReadyToUpgrade: 4,
+  NoUpdatesFound: 5,
+  WaitingForSync: 6,
+  Draining: 7,
+  Unknown: 8,
+};
+
+const TextButton = ({ title }) => (
   <Text style={styles.textButton}>{title}</Text>
 );
 
-const UpgradeBar = ({ upgradeInfo, onInstallPress }) => {
+const UpgradeBar = ({ peers }: { peers: Peer[] }) => {
   const { formatMessage: t } = useIntl();
+  const [upgradeInfo, setUpgradeInfo] = React.useState({});
+  const [backendState, setBackendState] = React.useState({});
+
+  // HACK(kira): I wanted the backend state tracking effect to be able to also
+  // trigger a re-render on an interval, which is what this state is for. I'm
+  // sure there's a nicer way to accomplish this.
+  const [fakeTrigger, setFakeTrigger] = React.useState(0);
+
+  // P2P Upgrades effects
+  // ----------------------------------------------------------------------
+  // Backend State + Peers => Frontend render effect
+  React.useEffect(() => {
+    if (!backendState.server) return;
+    const rState = getFrontendStateFromUpgradeState(backendState, peers);
+    setUpgradeInfo(rState);
+  }, [peers, backendState, fakeTrigger]);
+
+  // Backend state tracking effect. Interfaces with Node process for state &
+  // control.
+  React.useEffect(() => {
+    let iv;
+    function onState(state) {
+      log("GOT BACKEND STATE", JSON.stringify(state));
+      setBackendState(state);
+    }
+
+    function onError(err) {
+      log("BACKEND STATE ERROR", err.message);
+      setUpgradeInfo({
+        state: UpgradeState.GenericError,
+        context: err.message,
+      });
+    }
+
+    function onReady() {
+      rnBridge.channel.removeListener("p2p-upgrades-backend-ready", onReady);
+      log("backend says they are ready!");
+      rnBridge.channel.post("p2p-upgrades-frontend-ready");
+
+      rnBridge.channel.addListener("p2p-upgrade::error", onError);
+      rnBridge.channel.addListener("p2p-upgrade::state", onState);
+      iv = setInterval(() => {
+        setFakeTrigger(Math.random());
+      }, 3000);
+
+      rnBridge.channel.post("p2p-upgrade::get-state");
+      rnBridge.channel.post("p2p-upgrade::start-services");
+    }
+    log("startup", upgradeInfo);
+    rnBridge.channel.addListener("p2p-upgrades-backend-ready", onReady);
+    rnBridge.channel.post("p2p-upgrades-frontend-ready");
+    setUpgradeInfo({ state: UpgradeState.Searching });
+
+    return () => {
+      log("cleanup!");
+      if (iv) clearInterval(iv);
+      rnBridge.channel.removeListener("p2p-upgrade::state", onState);
+      rnBridge.channel.removeListener("p2p-upgrade::error", onError);
+      rnBridge.channel.post("p2p-upgrade::stop-services");
+    };
+  }, []);
+  // ----------------------------------------------------------------------
+
+  function onInstallPress() {
+    if (upgradeInfo.state !== UpgradeState.ReadyToUpgrade) {
+      log("### UPGRADE: not in ready state");
+      return;
+    }
+    ApkInstaller.install(upgradeInfo.context.filename)
+      .then(() => {
+        log("### UPGRADE: install ok!");
+      })
+      .catch(err => {
+        log("### UPGRADE: apk install failed", err);
+      });
+  }
+
   return (
     <TouchableNativeFeedback>
       <View style={styles.upgradeBar}>
@@ -268,6 +365,93 @@ const UpgradeBar = ({ upgradeInfo, onInstallPress }) => {
     </TouchableNativeFeedback>
   );
 };
+
+function getFrontendStateFromUpgradeState(state, peers) {
+  // Error present in ANY subsystem.
+  if (state.server.state === BackendUpgradeState.Server.Error) {
+    return {
+      state: UpgradeState.GenericError,
+      context: state.server.context.message,
+    };
+  }
+  if (state.downloader.search.state === BackendUpgradeState.Search.Error) {
+    return {
+      state: UpgradeState.GenericError,
+      context: state.downloader.search.context.message,
+    };
+  }
+  if (state.downloader.download.state === BackendUpgradeState.Download.Error) {
+    return {
+      state: UpgradeState.GenericError,
+      context: state.downloader.download.context.message,
+    };
+  }
+  if (state.downloader.check.state === BackendUpgradeState.Check.Error) {
+    return {
+      state: UpgradeState.GenericError,
+      context: state.downloader.check.context.message,
+    };
+  }
+
+  // Edge case I've (kira) seen once that shouldn't happen.
+  if (state.downloader.search.state === BackendUpgradeState.Search.Idle) {
+    return {
+      state: UpgradeState.GenericError,
+      context: "Search did not initialize correctly.",
+    };
+  }
+
+  // Upgrade available + waiting for syncs to finish.
+  if (
+    state.downloader.check.state === BackendUpgradeState.Check.Available &&
+    peers.some(p => p.status === peerStatus.PROGRESS)
+  ) {
+    return { state: UpgradeState.WaitingForSync, context: null };
+  }
+
+  // Upgrade available + not waiting for syncs to finish.
+  if (state.downloader.check.state === BackendUpgradeState.Check.Available) {
+    return {
+      state: UpgradeState.ReadyToUpgrade,
+      context: state.downloader.check.context,
+    };
+  }
+
+  // Upgrade is downloading.
+  if (
+    state.downloader.download.state === BackendUpgradeState.Download.Downloading
+  ) {
+    const progress = state.downloader.download.context;
+    return {
+      state: UpgradeState.Downloading,
+      context: { progress: progress.sofar / progress.total },
+    };
+  }
+
+  // Subsystem has been searching for upgrades for < 14 seconds.
+  if (
+    state.downloader.search.state === BackendUpgradeState.Search.Searching &&
+    Date.now() - state.downloader.search.context.startTime < 14 * 1000
+  ) {
+    return { state: UpgradeState.Searching, context: null };
+  }
+
+  // Subsystem is still uploading an upgrade to other peers.
+  if (
+    (state.server.state === BackendUpgradeState.Server.Sharing ||
+      state.server.state === BackendUpgradeState.Server.Draining) &&
+    state.server.context.length > 0
+  ) {
+    return { state: UpgradeState.Draining, context: null };
+  }
+
+  // Subsystem has been searching for upgrades for > 14 seconds.
+  if (state.downloader.search.state === BackendUpgradeState.Search.Searching) {
+    return { state: UpgradeState.NoUpdatesFound, context: null };
+  }
+
+  return { state: UpgradeState.Unknown, context: null };
+}
 
 export default UpgradeBar;
